refactor(Viewer): Move viewers by file in a specific folder
no code change. The style is split between root dans ./ViewersByFile folder
