+++ react/Viewer/ViewerByFile.jsx
@@ -6,15 +6,15 @@ import { FileDoctype } from '../proptypes'
 import withBreakpoints from '../helpers/withBreakpoints'
 
 import { isPlainText } from './helpers'
-import ImageViewer from './ImageViewer'
-import AudioViewer from './AudioViewer'
-import VideoViewer from './VideoViewer'
-import PdfJsViewer from './PdfJsViewer'
-import TextViewer from './TextViewer'
-import PdfMobileViewer from './PdfMobileViewer'
+import ImageViewer from './ViewersByFile/ImageViewer'
+import AudioViewer from './ViewersByFile/AudioViewer'
+import VideoViewer from './ViewersByFile/VideoViewer'
+import PdfJsViewer from './ViewersByFile/PdfJsViewer'
+import TextViewer from './ViewersByFile/TextViewer'
+import PdfMobileViewer from './ViewersByFile/PdfMobileViewer'
 import NoViewer from './NoViewer'
-import ShortcutViewer from './ShortcutViewer'
-import OnlyOfficeViewer from './OnlyOfficeViewer'
+import ShortcutViewer from './ViewersByFile/ShortcutViewer'
+import OnlyOfficeViewer from './ViewersByFile/OnlyOfficeViewer'
 
 import { useEncrypted } from './EncryptedProvider'
