+++ react/Viewer/ViewerByFile.spec.jsx
@@ -6,15 +6,15 @@ jest.mock('cozy-device-helper', () => ({
   ...jest.requireActual('cozy-device-helper'),
   isMobile: jest.fn()
 }))
-jest.mock('./ShortcutViewer', () => 'ShortcutViewer')
-jest.mock('./ImageViewer', () => 'ImageViewer')
-jest.mock('./AudioViewer', () => 'AudioViewer')
-jest.mock('./AudioViewer', () => 'AudioViewer')
-jest.mock('./VideoViewer', () => 'VideoViewer')
-jest.mock('./PdfJsViewer', () => 'PdfJsViewer')
-jest.mock('./PdfMobileViewer', () => 'PdfMobileViewer')
-jest.mock('./TextViewer', () => 'TextViewer')
-jest.mock('./OnlyOfficeViewer', () => 'OnlyOfficeViewer')
+jest.mock('./ViewersByFile/ShortcutViewer', () => 'ShortcutViewer')
+jest.mock('./ViewersByFile/ImageViewer', () => 'ImageViewer')
+jest.mock('./ViewersByFile/AudioViewer', () => 'AudioViewer')
+jest.mock('./ViewersByFile/AudioViewer', () => 'AudioViewer')
+jest.mock('./ViewersByFile/VideoViewer', () => 'VideoViewer')
+jest.mock('./ViewersByFile/PdfJsViewer', () => 'PdfJsViewer')
+jest.mock('./ViewersByFile/PdfMobileViewer', () => 'PdfMobileViewer')
+jest.mock('./ViewersByFile/TextViewer', () => 'TextViewer')
+jest.mock('./ViewersByFile/OnlyOfficeViewer', () => 'OnlyOfficeViewer')
 jest.mock('./NoViewer', () => 'NoViewer')
 
 describe('getViewerComponentName', () => {
