+++ react/Viewer/ViewersByFile/PdfJsViewer.spec.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
+
 import { PdfJsViewer, MIN_SCALE, MAX_SCALE, MAX_PAGES } from './PdfJsViewer'
 
 describe('PDFViewer', () => {
