+++ react/Viewer/ViewersByFile/styles.styl
@@ -1,0 +1,92 @@
+@require 'settings/breakpoints.styl'
+@require 'settings/palette.styl'
+@require 'settings/z-index.styl'
+@require '../controls.styl'
+
+$viewerHeightMedium = $toolbarHeightMedium + $footerHeight
+$viewerMarginTopMedium = $toolbarHeightMedium - $footerHeight
+
+.viewer-imageviewer
+.viewer-noviewer
+.viewer-audioviewer
+.viewer-videoviewer
+.viewer-pdfviewer
+.viewer-textviewer
+.viewer-canceled
+    position         relative
+    display          flex
+    width            60%
+    max-height       'calc(100% - %s)' % $toolbarHeight
+    justify-content  center
+    align-items      center
+    flex-direction   column
+    margin-top       $toolbarHeight
+
+    h2
+        max-width   90%
+
+    +medium-screen()
+        margin-left      0
+        width            100%
+        max-height       'calc(100% - %s)' % $viewerHeightMedium
+        margin-top       $viewerMarginTopMedium
+
+// rules for specific viewers below
+
+.viewer-videoviewer
+    video
+        width       100%
+        max-width   31.25rem
+        height      auto
+
+.viewer-textviewer
+    .viewer-textviewer-content
+        white-space pre-line
+        width 100%
+        max-height 70%
+        overflow auto
+
+    a
+        color var(--azure)
+
+    +medium-screen()
+        width 90%
+        margin-left 5%
+
+.viewer-filename
+    max-width     90%
+    text-overflow ellipsis
+    overflow hidden
+
+.viewer-pdfviewer-pdf
+    overflow auto
+    width 100%
+
+.viewer-pdfviewer-page > *
+    margin auto
+
+.viewer-pdfviewer-toolbar
+    position absolute
+    bottom 2rem
+    background var(--charcoalGrey)
+    color var(--white)
+    border-radius .5rem
+
+.viewer-imageviewer
+    flex             1 1 100%
+    width            auto
+    max-width        100%
+
+    img
+        display     block
+        max-width   100%
+        max-height  100%
+        box-shadow  0 .375rem 1.5rem 0 rgba(0, 0, 0, .5)
+
+.viewer-pdfMobile
+    width 100%
+    height 'calc(100% - %s - var(--flagship-top-height))' % $viewerHeightMedium
+    margin-top 'calc(var(--flagship-top-height) + %s)' % $viewerMarginTopMedium
+
+    &--image
+        width 100%
