+++ react/Viewer/withFileUrl.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 
 import ViewerSpinner from './ViewerSpinner'
-import NoNetworkViewer from './NoNetworkViewer'
+import NoNetworkViewer from './ViewersByFile/NoNetworkViewer'
 import { isFileEncrypted } from './helpers'
 
 const TTL = 6000
